Rename Frame to SchedulerLogicState (and other associated renames)
The concept of frame was too muddy. Frame is specifically being defined as the
state of the schedulerLogic between each handling of an incoming event. The
frame state evolves as we handle various pieces of logic in response to that
incoming event.

We also double-down on separating, even though it is just semantically, the
difference between intents and events. Intents are things not yet applied.
Events are things that were applied and we're notifying you about. StateEvents
can be returned as an intent to be applied (since per the event-sourcing
pattern we evolve our state via the same events that we emit and persist), but
after they are applied, they're referred to as events and not intents.