Introduce `OfferMatcher` interface (#73)
Summary:
we extract the offer matching portion of the `MesosEventsLogic` behind an `OfferMatcher` interface separating it from the rest of the logic e.g. building `TaskInfo`s. The only method of the interface receives a Mesos offer and a set of PodSpecs and returns a map of all matched PodSpecs with the list of corresponding resources. This will allow us to easier test different offer matching strategies in the future.