Eager cancel the graph fanout (#75)
Prior to this change, the closing of one of the outputs would not affect the
other. With this change, if the Mesos client closes, then the stream for the scheduler
state events will also be closed. If the scheduler state events stream is closed, then the Mesos client is closed.