Imperative SchedulerCommand interface
We update the SchedulerLogic interace to receive LaunchCommands rather
than the use the declarative replication scheme. In the end, this is
simpler and fits the problem domain better.

We remove SpecsSnapshot concern, and clean up quite a bit of confusing
code in the process.

Docs are updated (both code and docs folder)