Introduce and use version script
This version script aids with auto-versionining USI by counting the number of commits to master since some point in the history.

Commits not-yet-merged to master are versioned by counting the number of
commits since their merge-base (the point of deviation); further, commits
not-yet-merged are treated as snapshot. If origin/master is at version 0.1.15,
and tharper/some-new-feature deviated two commits before origin/master, it's
version will be 0.1.13-deadbee-SNAPSHOT.