Publish USI Binaries to Mesosphere Maven Repo (#100)
Summary:
This patch introduces a Jenkins job that publishes the USI binaries to our own
Maven repository. It uses the same credentials as the Marathon JAR publishing
job.

We had to for Jenkins since the credentials in Travis CI are not allowed to upload
artifacts to our S3 Maven bucket.