Multi-role, debounced, robust suppress and revive
We introduce a debounced suppress-and-revive mechanism which properly supports multiple Mesos roles (via the
UpdateFramework calls).

Further, we introduce the factory pattern to decouple stream wiring logic from component initialization and respective
dependencies.

Also, a generated-from-code mechanism is added to generate the visualization docs/scheduler-stream.svg; since this is
generated from code, it will not be wrong. The graph is updated when the unit tests are run.