Replace ask with sending request to make Mesos calls. (#151)
Summary:
The last MWT showed `AskTimeoutExceptions` when Mesos would not respond in time during
a Mesos master failover. This timeout was the result of an HTTP request timing out and is hard
to debug. It also leaves dead letters.

This change sends a promise instead. The promise is completed with the HTTP request result
which can include a timeout. It should ease debugging.